requirements.txt
flask
Flask-SQLAlchemy
Flask-Limiter
Flask-WTF
python-dateutil
pypinyin
Pillow
requests
git+ssh://git@github.com/patakk/hanzipy.git